refactor(gaussianSmooth): consolidate worker into paint tools directory
- Move gaussianSmooth.worker.ts to src/core/tools/paint/ directory
- Remove standalone gaussianSmooth.ts wrapper
- Integrate worker management directly into store
- Maintain consistent architecture with other paint tools
