+++ src/store/tools/gaussianSmooth.ts
@@ -1,15 +1,55 @@
 import { defineStore } from 'pinia';
 import { ref } from 'vue';
 import { TypedArray } from '@kitware/vtk.js/types';
+import * as Comlink from 'comlink';
 import vtkLabelMap from '@/src/vtk/LabelMap';
-import {
-  gaussianSmoothLabelMap,
-  type GaussianSmoothParams,
-} from '@/src/core/tools/gaussianSmooth';
+import { gaussianSmoothLabelMapWorker } from '@/src/core/tools/paint/gaussianSmooth.worker';
 
 const DEFAULT_SIGMA = 1.0;
 const MIN_SIGMA = 0.1;
 const MAX_SIGMA = 5.0;
+
+// Worker management
+type WorkerApi = {
+  gaussianSmoothLabelMapWorker: typeof gaussianSmoothLabelMapWorker;
+};
+
+let workerInstance: Comlink.Remote<WorkerApi> | null = null;
+
+async function getWorker() {
+  if (!workerInstance) {
+    // Set up worker with Comlink
+    const worker = new Worker(
+      new URL(
+        '@/src/core/tools/paint/gaussianSmooth.worker.ts',
+        import.meta.url
+      ),
+      { type: 'module' }
+    );
+    workerInstance = Comlink.wrap<WorkerApi>(worker);
+  }
+  return workerInstance;
+}
+
+async function gaussianSmoothLabelMap(
+  labelMap: vtkLabelMap,
+  params: { sigma: number; label: number }
+) {
+  const scalars = labelMap.getPointData().getScalars();
+  const originalData = scalars.getData();
+  const dimensions = labelMap.getDimensions();
+
+  const worker = await getWorker();
+
+  // Transfer data to worker
+  const workerInput = {
+    data: originalData,
+    dimensions,
+    params,
+  };
+
+  return worker.gaussianSmoothLabelMapWorker(workerInput);
+}
 
 export const useGaussianSmoothStore = defineStore('gaussianSmooth', () => {
   const sigma = ref(DEFAULT_SIGMA);
@@ -22,7 +62,7 @@ export const useGaussianSmoothStore = defineStore('gaussianSmooth', () => {
     segImage: vtkLabelMap,
     activeSegment: number
   ): Promise<TypedArray> {
-    const params: GaussianSmoothParams = {
+    const params = {
       sigma: sigma.value,
       label: activeSegment,
     };
